project_3_chips_defects_semantic_segmentation/requirements.txt
imageio
matplotlib
numba
numpy
opencv-python
Pillow
scikit-image
tifffile
tqdm